+++ capstone/SStream.c
@@ -6,12 +6,14 @@
 #endif
 #include <stdarg.h>
 #if defined(CAPSTONE_HAS_OSXKERNEL)
+#include <Availability.h>
 #include <libkern/libkern.h>
+#include <i386/limits.h>
 #else
 #include <stdio.h>
+#include <limits.h>
 #endif
 #include <string.h>
-#include <limits.h>
 
 #include <platform.h>
 
@@ -56,10 +58,6 @@ void SStream_concat(SStream *ss, const char *fmt, ...)
 // print number with prefix #
 void printInt64Bang(SStream *O, int64_t val)
 {
-	if (val == LONG_MIN) {
-		return;
-	}
-
 	if (val >= 0) {
 		if (val > HEX_THRESHOLD)
 			SStream_concat(O, "#0x%"PRIx64, val);
@@ -67,7 +65,10 @@ void printInt64Bang(SStream *O, int64_t val)
 			SStream_concat(O, "#%"PRIu64, val);
 	} else {
 		if (val <- HEX_THRESHOLD)
-			SStream_concat(O, "#-0x%"PRIx64, -val);
+			if (val == LONG_MIN)
+				SStream_concat(O, "#-0x%"PRIx64, val);
+			else
+				SStream_concat(O, "#-0x%"PRIx64, -val);
 		else
 			SStream_concat(O, "#-%"PRIu64, -val);
 	}
@@ -84,10 +85,6 @@ void printUInt64Bang(SStream *O, uint64_t val)
 // print number
 void printInt64(SStream *O, int64_t val)
 {
-	if (val == LONG_MIN) {
-		return;
-	}
-
 	if (val >= 0) {
 		if (val > HEX_THRESHOLD)
 			SStream_concat(O, "0x%"PRIx64, val);
@@ -95,7 +92,10 @@ void printInt64(SStream *O, int64_t val)
 			SStream_concat(O, "%"PRIu64, val);
 	} else {
 		if (val <- HEX_THRESHOLD)
-			SStream_concat(O, "-0x%"PRIx64, -val);
+			if (val == LONG_MIN)
+				SStream_concat(O, "-0x%"PRIx64, val);
+			else
+				SStream_concat(O, "-0x%"PRIx64, -val);
 		else
 			SStream_concat(O, "-%"PRIu64, -val);
 	}
@@ -104,22 +104,17 @@ void printInt64(SStream *O, int64_t val)
 // print number in decimal mode
 void printInt32BangDec(SStream *O, int32_t val)
 {
-	if (val == INT_MIN) {
-		return;
-	}
-
 	if (val >= 0)
 		SStream_concat(O, "#%u", val);
 	else
-		SStream_concat(O, "#-%u", -val);
+		if (val == INT_MIN)
+			SStream_concat(O, "#-%u", val);
+		else
+			SStream_concat(O, "#-%u", -val);
 }
 
 void printInt32Bang(SStream *O, int32_t val)
 {
-	if (val == INT_MIN) {
-		return;
-	}
-
 	if (val >= 0) {
 		if (val > HEX_THRESHOLD)
 			SStream_concat(O, "#0x%x", val);
@@ -127,7 +122,10 @@ void printInt32Bang(SStream *O, int32_t val)
 			SStream_concat(O, "#%u", val);
 	} else {
 		if (val <- HEX_THRESHOLD)
-			SStream_concat(O, "#-0x%x", -val);
+			if (val == INT_MIN)
+				SStream_concat(O, "#-0x%x", val);
+			else
+				SStream_concat(O, "#-0x%x", -val);
 		else
 			SStream_concat(O, "#-%u", -val);
 	}
@@ -135,10 +133,6 @@ void printInt32Bang(SStream *O, int32_t val)
 
 void printInt32(SStream *O, int32_t val)
 {
-	if (val == INT_MIN) {
-		return;
-	}
-
 	if (val >= 0) {
 		if (val > HEX_THRESHOLD)
 			SStream_concat(O, "0x%x", val);
@@ -146,7 +140,10 @@ void printInt32(SStream *O, int32_t val)
 			SStream_concat(O, "%u", val);
 	} else {
 		if (val <- HEX_THRESHOLD)
-			SStream_concat(O, "-0x%x", -val);
+			if (val == INT_MIN)
+				SStream_concat(O, "-0x%x", val);
+			else
+				SStream_concat(O, "-0x%x", -val);
 		else
 			SStream_concat(O, "-%u", -val);
 	}