+++ capstone/arch/X86/X86ATTInstPrinter.c
@@ -23,6 +23,7 @@
 #endif
 #include <platform.h>
 #if defined(CAPSTONE_HAS_OSXKERNEL)
+#include <Availability.h>
 #include <libkern/libkern.h>
 #else
 #include <stdio.h>
@@ -640,17 +641,7 @@ static void printMemReference(MCInst *MI, unsigned Op, SStream *O)
 			MI->flat_insn->detail->x86.operands[MI->flat_insn->detail->x86.op_count].mem.disp = DispVal;
 		if (DispVal) {
 			if (MCOperand_getReg(IndexReg) || MCOperand_getReg(BaseReg)) {
-				if (DispVal < 0) {
-					if (DispVal <  -HEX_THRESHOLD)
-						SStream_concat(O, "-0x%"PRIx64, -DispVal);
-					else
-						SStream_concat(O, "-%"PRIu64, -DispVal);
-				} else {
-					if (DispVal > HEX_THRESHOLD)
-						SStream_concat(O, "0x%"PRIx64, DispVal);
-					else
-						SStream_concat(O, "%"PRIu64, DispVal);
-				}
+				printInt64(O, DispVal);
 			} else {
 				// only immediate as address of memory
 				if (DispVal < 0) {