+++ capstone/arch/X86/X86DisassemblerDecoder.h
@@ -589,7 +589,7 @@ typedef struct InternalInstruction {
   uint8_t                       sib;
   /* The displacement, used for memory operands */
   bool                          consumedDisplacement;
-  int32_t                       displacement;
+  int64_t                       displacement;
   /* The value of the two-byte escape prefix (usually 0x0f) */
   uint8_t twoByteEscape;
   /* The value of the three-byte escape prefix (usually 0x38 or 0x3a) */
@@ -604,6 +604,11 @@ typedef struct InternalInstruction {
 
   /* The value of the vector extension prefix(EVEX/VEX/XOP), if present */
   uint8_t vectorExtensionPrefix[4];
+
+  /* Offsets from the start of the instruction to the pieces of data, which is
+     needed to find relocation entries for adding symbolic operands */
+  uint8_t displacementOffset;
+  uint8_t immediateOffset;
 
   // end-of-zero-members
 
@@ -646,11 +651,6 @@ typedef struct InternalInstruction {
 
   uint8_t immSize;	// immediate size for X86_OP_IMM operand
 
-  /* Offsets from the start of the instruction to the pieces of data, which is
-     needed to find relocation entries for adding symbolic operands */
-  uint8_t displacementOffset;
-  uint8_t immediateOffset;
-
   /* opcode state */
 
   /* The last byte of the opcode, not counting any ModR/M extension */
@@ -679,6 +679,9 @@ typedef struct InternalInstruction {
   /* The ModR/M byte, which contains most register operands and some portion of
      all memory operands */
   uint8_t                       modRM;
+
+  /* contains the location (for use with the reader) of the modRM byte */
+  uint64_t                      modRMLocation;
 
   // special data to handle MOVcr, MOVdr, MOVrc, MOVrd
   uint8_t                       firstByte;     // save the first byte in stream