+++ CHANGELOG.md
@@ -4,6 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+### Added
+* TODO (or remove section if none)
+
+### Changed
+- Upgraded bundled capstone from
+  [7e004bd4 to a31b5328](https://github.com/aquynh/capstone/compare/7e004bd4...a31b5328),
+  which incorporates upstream Capstone PR
+  [#1171](https://github.com/aquynh/capstone/pull/1171)
+
+### Deprecated
+- TODO (or remove section if none)
+
+### Removed
+- TODO (or remove section if none)
+
+### Fixed
+- TODO (or remove section if none)
+
+### Security
+- TODO (or remove section if none)
+
+
 ## [0.8.0] - 2018-06-02
 ### Added
 - Documented FreeBSD support