+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+### Added
+* TODO (or remove section if none)
+
+### Changed
+- Upgraded bundled capstone to release [3.0.5](https://github.com/aquynh/capstone/releases/tag/3.0.5)
+  (Git commit [a31b5328 to db19431d](https://github.com/aquynh/capstone/compare/a31b5328...db19431d)).
+
+### Deprecated
+- TODO (or remove section if none)
+
+### Removed
+- TODO (or remove section if none)
+
+### Fixed
+- TODO (or remove section if none)
+
+### Security
+- TODO (or remove section if none)
+
+
 ## [0.9.0] - 2018-07-08
 
 ### Changed