+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - YYYY-MM-DD
+
+### Changed
+- Bump minimum Rust version to 1.24.1
+
+
 ## [0.5.0] - 2018-09-21
 ### Added
 * `InsnDetail` to preamble